refactor: Move Layout & Providers to specific files
